requirements.txt
django==3.1.2
psycopg2-binary==2.8.6
djangorestframework==3.12.4
django-cors-headers==3.0.0
gunicorn